Export XML fix

+++ include/DiscreteVerification/VerificationTypes/SMCVerification.hpp
@@ -47,7 +47,7 @@ class SMCVerification : public Verification<RealMarking> {
         SMCSettings smcSettings;
         size_t numberOfRuns;
         unsigned int maxTokensSeen;
-        unsigned long totalTime = 0;
+        double totalTime = 0;
         unsigned long totalSteps = 0;
         int64_t durationNs = 0;
